move TextInput into elements module, and delete helpers now that it's gone
