--- tuig-iosys/src/ui/helpers/mod.rs
@@ -1,4 +1,0 @@
-//! Miscellaneous IO-related helpers to package up common functionality.
-
-mod text_input;
-pub use text_input::{TextInput, TextInputRequest};
